Move AST/Context.h to Queries/Context.h
---
